Clean up app-core.js
